@@ -1,10 +1,10 @@
 {
   "name": "@dzangolab/vue-accordion",
-  "version": "1.2.0",
-  "description": "An accordion component for vue.js",
+  "version": "2.0.0-rc.1",
+  "description": "An accordion component for vue.js (vue3)",
   "author": {
     "name": "Olivier Pichon",
-    "email": "op@dzango.com",
+    "email": "info@dzango.com",
     "url": "https://github.com/opichon"
   },
   "scripts": {
@@ -27,18 +27,19 @@
     "@commitlint/config-conventional": "^13.2",
     "@fortawesome/fontawesome-svg-core": "^1.2",
     "@fortawesome/free-solid-svg-icons": "^5.15",
-    "@fortawesome/vue-fontawesome": "^2.0",
+    "@fortawesome/vue-fontawesome": "^3.0.0-4",
     "@vue/cli-plugin-babel": "^4.5",
     "@vue/cli-plugin-eslint": "^4.5",
     "@vue/cli-plugin-unit-jest": "^4.5",
     "@vue/cli-service": "^4.5",
-    "@vue/test-utils": "1.2",
+    "@vue/compiler-sfc": "^3.2.20",
+    "@vue/test-utils": "2.0.0-rc.15",
     "@vuedoc/md": "^3.2",
-    "@vuedoc/parser": "^3.4",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.0",
-    "babel-jest": "^27.2",
+    "babel-jest": "^26.0",
     "bootstrap": "^5.1",
+    "core-js": "^3.6.5",
     "eslint": "^7.32, < 8.0",
     "eslint-plugin-import": "^2.22",
     "eslint-plugin-node": "^11.1",
@@ -49,8 +50,8 @@
     "node-sass": "^6.0",
     "sass-loader": "^10, < 11",
     "standard-version": "^9.3",
-    "vue": "^2.6",
-    "vue-template-compiler": "^2.6"
+    "vue": "^3.2.20",
+    "vue-jest": "^5.0.0-0"
   },
   "bugs": "https://github.com/dzangolab/vue-accordion/issues",
   "files": [
