@@ -136,7 +136,7 @@ data() {
       @item:select="onItemSelect"
     >
       <template #title>
-        <!-- Use this lsot to customize each item's title -->
+        <!-- Use this slot to customize each item's title -->
         <slot
           name="title"
           :index="key"
@@ -145,7 +145,7 @@ data() {
       </template>
 
       <template #toggle>
-        <!-- Use this lsot to customize each item's toggle button -->
+        <!-- Use this slot to customize each item's toggle button -->
         <slot name="toggle" :active="isActive(key)" />
       </template>
 
